[Maint] whitespace and test fixing
