+++ .gitignore
@@ -191,5 +191,5 @@ Thumbs.db
 .pixi
 pixi.lock
 
-#vscode git/repo/workspace config 
+#vscode git/repo/workspace config
 .vscode/
