+++ tests/integration/test_pipeline.py
